EES-1259 Merge importService into releaseDataFileService
